Made the variables advanced
Fixed the program so that it pulled the variables from a CSV not an internal dict.

Changed the data structure from the single "ds" dictionary to 3 (geo, var and variables)

Fixed various bugs that showed up along the way

Next steps is to play with group variables and the "by" column in the spreadsheet for more detailed variables like age and income.